requirements.txt
asgiref==3.8.1
Django==5.2.4
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-model-utils==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
pillow==11.3.0
psycopg==3.2.9
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.14.1
ruff==0.11.8
